requirements.txt
absl-py==0.7.1
appdirs==1.4.3
appnope==0.1.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cmake
colorama==0.4.1
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-js-asset==1.2.2
django-paypal==1.0.0
djangorestframework==3.9.2
dlib
docutils==0.14
entrypoints==0.3
face-recognition==1.2.3
face-recognition-models==0.3.0
future==0.17.1
gast==0.2.2
grpcio==1.21.1
gunicorn
h5py==2.9.0
idna==2.8
imdirect==0.5.0
imutils==0.5.2
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
isort==4.3.13
jedi==0.14.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.4.1
notebook==5.7.8
numpy==1.16.3
oauthlib==3.0.2
opencv-python==4.1.0.25
opencv-utils==0.0.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==5.4.1
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2==2.7.7
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
PyInquirer==1.0.3
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.0
pyrsistent==0.15.2
PySocks==1.7.0
python-dateutil==2.8.0
python3-openid==3.1.0
pyttsx3==2.7
pytz==2018.9
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.1
readme-renderer==24.0
regex==2019.6.8
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rply==0.7.7
scikit-learn==0.21.2
scipy==1.3.0
selenium==3.141.0
Send2Trash==1.5.0
serial==0.0.97
six==1.12.0
sklearn==0.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
tqdm==4.32.2
traitlets==4.3.2
twilio==6.29.3
twine==1.13.0
typed-ast==1.3.1
urllib3==1.25.3
uuid==1.30
virtualenv==16.6.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whitenoise==4.1.2
widgetsnbextension==3.4.2
wrapt==1.11.1
